# server/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is Accessed Job Service Documentation",
    "version": "1.0.0",
    "title": "Accessed Job Service",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "accessed-job-service.us-west-2.elasticbeanstalk.com",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/jobs/search": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Add a new job to the company",
        "description": "",
        "operationId": "addJob",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Job object that needs to be added to the company",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JobRequisition"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update an existing job - In-Progress",
        "description": "",
        "operationId": "updateJob",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Job object that needs to be added to the company",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JobRequisition"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Job not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Search existing jobs",
        "description": "",
        "operationId": "getJobs",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Job not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Job By Id",
        "description": "Get Job By Id",
        "operationId": "getJobById",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "ID of job to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Job not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "accessed_auth": {
      "type": "basic"
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "JobRequisition": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "jobId": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "durationMonths": {
          "type": "integer"
        },
        "minMonthExperience": {
          "type": "integer"
        },
        "maxMonthExperience": {
          "type": "integer"
        },
        "lastCurrencyUom": {
          "type": "string"
        },
        "noOfResources": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "expirationDate": {
          "type": "integer"
        },
        "requiredOnDate": {
          "type": "integer"
        },
        "salaryRangeLow": {
          "type": "integer"
        },
        "salaryRangeHigh": {
          "type": "integer"
        },
        "salaryFixed": {
          "type": "integer"
        },
        "noApplied": {
          "type": "integer"
        },
        "level": {
          "type": "string"
        },
        "jobFunction": {
          "type": "object"
        },
        "responsibilities": {
          "type": "array",
          "items": {}
        },
        "qualifications": {
          "type": "array",
          "items": {}
        },
        "skills": {
          "type": "array",
          "items": {}
        },
        "createdDate": {
          "type": "integer"
        },
        "employmentType": {
          "type": "string"
        },
        "promotion": {
          "type": "object"
        },
        "isSaved": {
          "type": "boolean"
        },
        "company": {
          "type": "object"
        },
        "connection": {
          "type": "object"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      },
      "xml": {
        "name": "JobRequisition"
      }
    },
...
  }
}